+++ release_notes.md
@@ -7,6 +7,7 @@
 * Improve tag splitting (opened related [bug report](https://github.com/dotnet/diagnostics/issues/5935))
 * Add `[processtags]` config to attach target data like filename, etc.
 * Add `[hosttags]` config to attach server data like hostname, env-vars, etc.
+* Improved logging, added a console logger for service mode output
 * Show Windows permissions reminder when running service as user
 * README warning about provider-wildcard overhead and filtering
 * README clarification that Uptime is not "transparent" to Collectors